fix: Problem with creating lessons without solution files
fix for https://github.com/stackblitz/tutorialkit/issues/107
In the hasSolution function of TutorialStore, `this._lessons.solution` is an array with the structure [lessonName, [list of solution files]]. Instead of checking the length of the solution array. we have to check the length of the second index which is the list of solution files.